requirements.txt
opencv-python>=4.8.0.76
streamlit>=1.28.0
torch>=2.0.1
sahi>=0.11.4,<0.12.0
realesrgan>=0.3.0
gfpgan>=1.3.8
numpy>=1.24.3
Pillow>=10.0.0
basicsr>=1.4.2
gdown>=4.7.1
pybboxes>=0.1.6